Extraigo los tests de indicadores a un archivo separado
