+++ tests/cassetes/indicators/test_date_network_indicators_empty_catalog.yaml
