+++ tests/cassetes/indicators/test_field_indicators_on_full_catalog.yaml
