+++ tests/cassetes/indicators/test_field_indicators_on_min_catalog.yaml
