+++ tests/cassetes/indicators/test_format_indicators.yaml
