+++ tests/cassetes/indicators/test_generate_catalog_indicators.yaml
