+++ tests/cassetes/indicators/test_indicators_invalid_periodicity.yaml
