+++ tests/cassetes/indicators/test_indicators_missing_periodicity.yaml
