+++ tests/cassetes/indicators/test_last_updated_indicator_missing_issued_field.yaml
