+++ tests/cassetes/indicators/test_network_indicators.yaml
